// src/server/game/Enums.ts
export enum eScoreType {
	SIX_OF_KIND_1S = "Six 1",
	SIX_OF_KIND = "Six of a kind",
	FIVE_OF_KIND_1S = "Five 1",
	FIVE_OF_KIND = "Five of a kind",
	FOUR_OF_KIND_1S = "Four 1",
	FOUR_OF_KIND = "Four of a kind",
	THREE_PAIR = "Three pair",
	STRAIGHT = "Straight",
	THREE_OF_KIND_1S = "Three 1",
	THREE_OF_KIND = "Three of a kind",
	ONE = "One",
	FIVE = "Five"
}